Move dev contracts to v0.5, and port to solidity 0.5.0
