+++ evm/v0.5/contracts/dev/CoordinatorInterface.sol
@@ -1,4 +1,4 @@
-pragma solidity 0.4.24;
+pragma solidity 0.5.0;
 pragma experimental ABIEncoderV2;
 
 contract CoordinatorInterface {
