+++ evm/v0.5/contracts/dev/LinkEx.sol
@@ -1,4 +1,4 @@
-pragma solidity 0.4.24;
+pragma solidity 0.5.0;
 
 import "./LinkExInterface.sol";
 import "openzeppelin-solidity/contracts/math/SafeMath.sol";
