+++ evm/v0.5/contracts/dev/LinkExInterface.sol
@@ -1,4 +1,4 @@
-pragma solidity 0.4.24;
+pragma solidity 0.5.0;
 
 interface LinkExInterface {
   function currentRate() external view returns (uint256);
