+++ evm/v0.5/contracts/dev/OracleDev.sol
@@ -1,4 +1,4 @@
-pragma solidity 0.4.24;
+pragma solidity 0.5.0;
 
 import "../Oracle.sol";
 import "./LinkExInterface.sol";
